\name{SQLiter-package}
\alias{SQLiter-package}
\alias{SQLiter}
\docType{package}
\title{
A few handy functions for lite work with SQLite
}
\description{
A few handy functions for lite work with SQLite
}
\details{
\tabular{ll}{
Package: \tab SQLiter\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-06-17\cr
License: \tab Apache 2.0\cr
}

}
\author{
Bill -Chopper- Genné

Maintainer:developer@meaningful-solutions.com
}
\references{
%%~~ Literature or other references for background information ~~
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
%%~~ directory ~~
\keyword{ package }
\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%~~ simple examples of the most important functions ~~
}
